Move project to accounts app
