+++ ployst/core/accounts/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Team
+from .models import Project, ProjectManager, Team
 
 
 class TeamUserInline(admin.TabularInline):
@@ -11,3 +11,13 @@ class TeamAdmin(admin.ModelAdmin):
     inlines = (TeamUserInline,)
 
 admin.site.register(Team, TeamAdmin)
+
+
+class ProjectManagerInline(admin.TabularInline):
+    model = ProjectManager
+
+
+class ProjectAdmin(admin.ModelAdmin):
+    inlines = (ProjectManagerInline,)
+
+admin.site.register(Project, ProjectAdmin)
