+++ ployst/core/accounts/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from .models import Project, ProjectManager
+
+
+class ProjectManagerSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ProjectManager
+
+
+class ProjectSerializer(serializers.ModelSerializer):
+    managers = ProjectManagerSerializer(many=True)
+
+    class Meta:
+        model = Project
