+++ ployst/core/accounts/urls.py
@@ -4,6 +4,7 @@ from rest_framework import routers
 from . import views
 
 router = routers.SimpleRouter()
+router.register(r'project', views.ProjectViewSet)
 router.register(r'team', views.TeamViewSet)
 
 urlpatterns = patterns(
