+++ ployst/core/accounts/views.py
@@ -1,6 +1,13 @@
 from rest_framework.viewsets import ModelViewSet
 
-from .models import Team
+from .models import Project, Team
+from .serializers import ProjectSerializer
+
+
+class ProjectViewSet(ModelViewSet):
+    model = Project
+    serializer_class = ProjectSerializer
+    filter_fields = ('team',)
 
 
 class TeamViewSet(ModelViewSet):
