+++ ployst/core/features/admin.py
@@ -1,19 +1,9 @@
 from django.contrib import admin
 
-from .models import Feature, Project, ProjectManager
+from .models import Feature
 
 
 class FeatureAdmin(admin.ModelAdmin):
     pass
 
 admin.site.register(Feature, FeatureAdmin)
-
-
-class ProjectManagerInline(admin.TabularInline):
-    model = ProjectManager
-
-
-class ProjectAdmin(admin.ModelAdmin):
-    inlines = (ProjectManagerInline,)
-
-admin.site.register(Project, ProjectAdmin)
