+++ ployst/core/features/models.py
@@ -1,36 +1,6 @@
-from django.contrib.auth.models import User
 from django.db import models
 
-from ..accounts.models import Team, TeamObject
-
-
-class Project(TeamObject):
-    """
-    A project groups features together.
-
-    Projects belong to teams.
-
-    """
-    name = models.CharField(max_length=100)
-    url = models.URLField()
-    team = models.ForeignKey(Team, related_name='projects')
-
-    team_lookup = 'team'
-
-    def __unicode__(self):
-        return self.name
-
-
-class ProjectManager(models.Model):
-    """
-    Users that can manage a project (besides the team managers).
-
-    These can manage project provider settings, but not user permissions,
-    which remain the realm of the team managers.
-
-    """
-    user = models.ForeignKey(User, related_name='managed_projects')
-    project = models.ForeignKey(Project, related_name='managers')
+from ..accounts.models import Project, TeamObject
 
 
 class Feature(TeamObject):
