+++ ployst/core/features/urls.py
@@ -4,7 +4,6 @@ from rest_framework import routers
 from . import views
 
 router = routers.SimpleRouter()
-router.register(r'project', views.ProjectViewSet)
 router.register(r'feature', views.FeatureViewSet)
 
 urlpatterns = patterns(
