+++ ployst/core/features/views.py
@@ -1,13 +1,7 @@
 from rest_framework.viewsets import ModelViewSet
 
-from .models import Feature, Project
-from .serializers import FeatureSerializer, ProjectSerializer
-
-
-class ProjectViewSet(ModelViewSet):
-    model = Project
-    serializer_class = ProjectSerializer
-    filter_fields = ('team',)
+from .models import Feature
+from .serializers import FeatureSerializer
 
 
 class FeatureViewSet(ModelViewSet):
