+++ requirements/runtime.txt
@@ -4,7 +4,7 @@ django==1.6.1
 django-appconf==0.6         # provides graceful configuration of apps
 django-compressor==1.3
 django-filter==0.7          # provides generic filtering in API
-djangorestframework==2.3.5
+djangorestframework==2.3.8
 South==0.7.6
 
 # Git provider requirements
